refactor: [sync] reorganize parser function path
