+++ src/pages/Synchronization/Datasets/DatasetGlobalSettings.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from '@dhis2/prop-types'
 import SettingsTable from '../../../components/settings-table/settings-table'
 import { DataSetting } from '../../../constants/data-set-settings'
-import { parseValueBySettingType } from '../../../modules/dataset/parseValueBySettingType'
+import { parseValueBySettingType } from './parseValueBySettingType'
 
 const DatasetGlobalSettings = ({ settings, handleChange, disable }) => {
     const handleInputTable = e => {
