+++ src/pages/Synchronization/Datasets/parseValueBySettingType.js
